+++ spec/xml/xslt_versions_spec.rb
@@ -401,10 +401,3 @@ RSpec.describe "predicate-pattern dispatch beyond 96 templates (leptris/leptris#
   end
 end
 
-RSpec.describe "Element#to_xml expand_empty: true (libleptris 1.9.95, leptris/leptris#882)" do
-  it "emits <a></a> instead of <a/> for empty elements (libxml2 NO_EMPTY_TAGS parity)" do
-    el = Leptris::XML::Document.parse("<r><a/></r>").root.children.first
-    expect(el.to_xml).to eq("<a/>")
-    expect(el.to_xml(expand_empty: true)).to eq("<a></a>")
-  end
-end